add support for platform arg
with the introduction of the `platform` flag to `img`, we're now able to
set what's the base platform that the images that we `FROM` should use.

(see https://github.com/genuinetools/img/pull/217)

this allows one that has `FROM ubuntu:bionic` to use an armv7-based base
image when running with `--platform=arm`, and an arm64 when using
`--platform=arm64`.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>